mail: shit has escalated to Infinite Complexity
.---------------------.
| Why Am I Doing This |
'---------------------'
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||

We have Postfix, Dovecot and OpenLDAP.  Dovecot seems to be able to find
users -- that is, `doveadm user` and `doveadm auth login` seem to work.
Postfix seems to sort-of work. Next up is to make Postfix use LMTP to
actually deliver the mail, so that only Dovecot needs to write to the
Maildir. That one fails because Dovecot can't find the user. I suppose
it's now time to write one of these funky-looking queries somewhere in
the badly documented Dovecot configuration. Yes, I said it. That shit is
only documented in an example config in the Git repo. There's no up to
date documentation. Is anyone but me using that thing? Maybe I want
another IMAP/POP3 server. With a separate SASL thing.

Compared to this, volume-hacking in the SSL certs will be a walk in the
park. Finding postgrey, policyd-spf and opendkim will be easy as well.

But the LDAP shit, gah, someone please end the pain! What alternatives
do I have? A Postgres database maybe, so that the queries are at least
normal SQL. I'm pretty sure the tried-and-true /etc/passwd with some
virtual maps, is not going to fly in Docker.

And do I want my Linux systems themselves to authenticate against an
OpenLDAP, with some PAM fuckery or what have you? Seems like a recipe
for disaster if Docker ever goes belly-up: now you can't log in to fix
that shit.

Did I mention the logging situation in Docker is a nuclear clusterfuck?
What the fuck is up with expecting everything to send its logs to
stdout/stderr? Your fancy Rails app might work like this, but grown-up
software does one or more of four things: 1) log its shit to a file on
disk (have fun with logrotate); 2) log its shit to syslog (/dev/log); 3)
log its shit to journald; 4) log its shit to stdout. Of course not every
daemon has different gradations of having hardcoded these choices or
not, or to have implemented them or not. Dovecot does 2, 1 or 4; Postfix
does 2; OpenLDAP does 4. Where does that get you the logs?  In Docker's
JSON storage, great! Yes, I know about logging drivers and they mostly
suck. Now I get to run a Splunk/ELK/Graylog stack!