rsyslog: basic rsyslog that listens on udp/514 [*]
Also it has half-baked stuff for being run as a sidecar, but a sidecar
makes for a racy solution in this case, with /dev/log being there or not
yet when your important daemon starts, depending on the phase on the
moon. Also can't just create the socket file beforehand.

It really sucks to not have a proper init system. Docker feels like a
pile of hacks on top of hacks, on top of hacks.